pid_controller and communication with arm is working

+++ hps-c/include/interface.hpp
@@ -45,7 +45,7 @@ public:
     uint timeout_ms = 10;
 private:
     uint rows, cols;
-    float pos;
+    int pos;
     uint ganglion_id = 0;
     uint motor_id = 0;
     char inputstring[30];
